misc/HtmlTestSuite/results/test-1367.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>Shadow DOM Test: Upper-boundary encapsulation: shadow root's Selector APIs</title>
<link href="mailto:sgrekhov@unipro.ru" rel="author" title="Sergey G. Grekhov"/>
<link href="mailto:mfursov@unipro.ru" rel="author" title="Mikhail Fursov"/>
<link href="mailto:yutak@google.com" rel="author" title="Yuta Kitamura"/>
<link href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#upper-boundary-encapsulation" rel="help"/>
<meta content="Upper-boundary encapsulation: Nodes in a shadow tree must be accessible through selector APIs of the shadow root." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="../../../../html/resources/common.js"/>
</head>
<body>
<div id="log"/>
<script>
// Return a document containing the structure below:
//
// &lt;body&gt; - - - - - {shadow-root}
//   |                    |
//   |                    +-- &lt;p class="test-class" id="test-id"&gt;
//   |
//   +-- &lt;p class="test-class" id="test-id"&gt;
function createTestDocument() {
    var doc = document.implementation.createHTMLDocument('Test');
    var pHost = doc.createElement('p');
    pHost.className = 'test-class';
    pHost.id = 'test-id';
    doc.body.appendChild(pHost);
    var shadowRoot = doc.body.attachShadow({mode: 'open'});
    var pShadow = doc.createElement('p');
    pShadow.className = 'test-class';
    pShadow.id = 'test-id';
    shadowRoot.appendChild(pShadow);
    return {
        doc: doc,
        shadowRoot: shadowRoot,
        pHost: pHost,
        pShadow: pShadow
    };
}

test(function () {
    var documentObject = createTestDocument();
    var shadowRoot = documentObject.shadowRoot;
    var pShadow = documentObject.pShadow;
    assert_equals(shadowRoot.querySelector('p'), pShadow);
    assert_equals(shadowRoot.querySelector('.test-class'), pShadow);
    assert_equals(shadowRoot.querySelector('#test-id'), pShadow);
},
    'Elements in a shadow tree should be accessible from ' +
    'shadow root\'s querySelector() method.'
);

function assert_singleton_node_list(nodeList, expectedNode) {
    assert_equals(nodeList.length, 1);
    assert_equals(nodeList[0], expectedNode);
}

test(function () {
    var documentObject = createTestDocument();
    var shadowRoot = documentObject.shadowRoot;
    var pShadow = documentObject.pShadow;
    assert_singleton_node_list(shadowRoot.querySelectorAll('p'), pShadow);
    assert_singleton_node_list(shadowRoot.querySelectorAll('.test-class'),
                               pShadow);
    assert_singleton_node_list(shadowRoot.querySelectorAll('#test-id'),
                               pShadow);
},
    'Elements in a shadow tree should be accessible from ' +
    'shadow root\'s querySelectorAll() method.'
);
</script>
</body>
</html>
